+++ Sources/Schemas/AccountTypeaheadValue.swift
@@ -1,0 +1,38 @@
+import Foundation
+
+public struct AccountTypeaheadValue: Codable, Hashable, Sendable {
+    public let accountIds: [String]
+    public let value: String
+    /// Additional properties that are not explicitly defined in the schema
+    public let additionalProperties: [String: JSONValue]
+
+    public init(
+        accountIds: [String],
+        value: String,
+        additionalProperties: [String: JSONValue] = .init()
+    ) {
+        self.accountIds = accountIds
+        self.value = value
+        self.additionalProperties = additionalProperties
+    }
+
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        self.accountIds = try container.decode([String].self, forKey: .accountIds)
+        self.value = try container.decode(String.self, forKey: .value)
+        self.additionalProperties = try decoder.decodeAdditionalProperties(using: CodingKeys.self)
+    }
+
+    public func encode(to encoder: Encoder) throws -> Void {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try encoder.encodeAdditionalProperties(self.additionalProperties)
+        try container.encode(self.accountIds, forKey: .accountIds)
+        try container.encode(self.value, forKey: .value)
+    }
+
+    /// Keys for encoding/decoding struct properties.
+    enum CodingKeys: String, CodingKey, CaseIterable {
+        case accountIds = "account_ids"
+        case value
+    }
+}