+++ Sources/Schemas/NotificationEventsForAdHocEnum.swift
@@ -2,7 +2,7 @@ import Foundation
 
 /// Limited notification events for ad-hoc per-subject notifications.
 /// 
-/// Set per order or per tracking session for any directory entry contact.
+/// Set per order or per tracking session for any contact.
 /// Limited scope since it can go  outside the org, e.g. share order completion but probably not billing events
 public enum NotificationEventsForAdHocEnum: String, Codable, Hashable, CaseIterable, Sendable {
     case shippingOrderStaged = "shipping.order.staged"